complete_or_balance_kernel: fix complete kernel.
When the subtree isn't shaped like a square/cube, we need an "end quadrant" q
so that the completion loop respects the shape of the subtree.  This end
quadrant was not initialized correctly in the case when the input array was
length 0: the correct initialization was copied from the length > 0 case, and
it seems to fix the issue.