Initial commit of vtu cell data export logic
This is a bit of a mess currently but it works; p4est_vtk_write_header
has been hijacked to append the custom cell data to the tree/level/rank
cell data exported therein. I do not like this solution but I do not
want to mess too much with p4est's paradigm without discussing it with
the main developers.