Document p4est_vtk_context_t struct, add a member
This commit adds Doxygen documentation for the p4est_vtk_context_t
structure and adds the new member num_nodes which will hold the number
of nodes in the p4est vtk output. This will be used to assure the proper
number of point scalars/vectors are provided by the user.