Update documentation
This commit updates the Doxygen documentation; it explains how
p?est_vtk_write_{cell,point}_data now requires that the current
p?est_vtk_context structure be provided as the final of the variable
arguments. This is used to validate that the correct number of variable
arguments have been provided.