Add Daniel's peroidic example: it fails until we fix edge and corner transforms
Computing transforms was complex and error prone.

Now we directly build a list of all the copies we wish to omit and
compare against it.

The "weird" example in p8est_test_periodic was wrong: the weird edge and
weird face needed to include all of the copies that self identify,
including the same edge with multiple orientations.

We have to allow multiple orientations of the same edge in
p8est_connectivity_is_valid()