add new test for mesh
instead of the old test, where optical verification was achieved we now
verify that the mesh is bijective by finding neighbor in certain
direction and then looking for the neighbor of the obtained neighbor in
the opposite direction to see if we can recover the original quadrant.

The optical test is kept but will be neither built nor run.
TODO: create make target test_mesh_optical