Clean up some bool comparisons
In C we may check for == 0 or != 0 (all nonzero values are true).
We may not check booleans for == 1: this may yield a false negative.

Some problems remain in test_quadrants with functions a() == !b():
if a returs true with a value != 1 and b is false, the outcome is wrong.
It's ok in this p4est test since the functions behave consistently.