Add coordinate comparison and transformations.
It is insufficient just to transform nodes because nodes are restricted to being at valid corners of cells:
not all coordinates can be represented by nodes.

Likewise, the coordinates can be sorted in Morton order even without a level
attached, which is useful for canonicalizing points between trees, so
we add this capability.

Implement p4est_coordinates_compare

Fix implementation and document coordinate comparison.

Fix quadrant to coordinates