MPI IO: Forced padding and newline halo for pad string
We enforce to start and end the padding string with a newline character
each. This means that we need to pad always with at least 2 bytes because
otherwise the newline breaks in the text editor view of a file would be
inconsistent. Furthermore, we enforce on this way that each line in the
text editor view is divisble by P4EST_BYTE_DIV since we moved the newline
at the beginnig of the array metadata to the padding string.