autotools: revert to not using autoreconf
autoreconf parses the Makefile.am and greps for ACLOCAL_AMFLAGS.
This means autoreconf relios on the repository version of Makefile.am,
which must not be modified by bootstrap or any other configuration calls.
Now, this line contains configure variables to be replaced for the Makefile
in accordance with autoconf variable substitution.
Now autoreconf breaks but that's a hack anyway.
We choose not to need it.