MPI IO: Distinguish file error codes
We distinguish the file error code range of p4est and libsc. This
enables us to check if a given error code is a p4est, a libsc or an
invalid file error code. These checks are perfomed in assertions and
such a range check is also used to directly copy file error codes that
are alread translated from libsc to p4est.