p4est_search: fix to pass true quadrant
In the relatively new function p4est_search_reorder, we accept
a parameter to switch level skipping on or off as an optimization.
If on and only a single leaf is contained in a search quadrant,
we directly skip down to the leaf as next quadrant to consider.

We fix the bug that the quadrant was not real when skipping was off.