Fix balance
In balance neighborhood_owned skips tree boundary quadrants that
are critical for balancing because they intersect the insulation layer
of quadrants in edge-neighboring trees. Therefore,
neighborhood_boundary_owned is introduced as a function that
determines always the full tree boundary if the tree is fully owned.
This means that in tree_compute_overlap more quadrants than necessary
are considered but this ensures correct results (cf. simple3 example drop).