ci:cmake: increase test timeouts to avoid transient failures
the runtime of tests on GitHub Actions and CI systems in general
has considerable variance due to different virtual worker hardware
and system loads.

The individual CTest test timeouts increase from 10 => 60 seconds.
The overall job time limit increases to 60 minutes. The only reason
for the overall time limit is so that a hung download or MPI job
doesn't sit "running" for a couple days.

also, include GCC-13 in tests.