p4est_bits: allow for coordinates on high faces
Previously, coordinates were treated like from the lower left corner of a
quadrant.  Thus, they were not allowed to touch any high face of the unit
tree.  We change this to make coordinates valid on all tree boundaries.  We
also allow coordinates as deep as P4EST_MAXLEVEL (quadrant midpoints).

The idea is to use the coordinates to index any boundary object of a
quadrant, thus any of its corners and face, edge, and volume midpoints.